InteractionMechanicUpdate
Now we have boxes that you can interact, when interact the box give you a "ingredient", that you can hold. We also have a table that you can interact too, when interact you put the ingredient on the table, if interact again you hold the item again. Note: the sprites is only there for examples